added fuel usage search

+++ sophena/src/sophena/rcp/editors/basedata/fuels/FuelEditor.java
@@ -20,11 +20,14 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import sophena.db.daos.FuelDao;
+import sophena.db.usage.SearchResult;
+import sophena.db.usage.UsageSearch;
 import sophena.model.Fuel;
 import sophena.rcp.App;
 import sophena.rcp.Images;
 import sophena.rcp.M;
 import sophena.rcp.editors.Editor;
+import sophena.rcp.editors.basedata.UsageError;
 import sophena.rcp.utils.Actions;
 import sophena.rcp.utils.Editors;
 import sophena.rcp.utils.KeyEditorInput;
@@ -155,6 +158,11 @@ public class FuelEditor extends Editor {
 					"Soll der ausgewählte Brennstoff wirklich gelöscht werden?");
 			if (!doIt)
 				return;
+			List<SearchResult> usage = new UsageSearch(App.getDb()).of(f);
+			if (!usage.isEmpty()) {
+				UsageError.show(usage);
+				return;
+			}
 			try {
 				dao.delete(f);
 				list.remove(f);
